Enable passing flag to CSS validator regarding vendor extensions
Allows calling set_vendor_extension_warning! on a validator instance
to have the CSS evaluated with vendor extensions considered as
warnings rather than errors.

Like so:
set_vendor_extension_warning!('Warnings')

will cause CSS like -moz-border-radius: 3px to be a warning not an error

The default has not changed (which is to say, use the W3C's default).

You can set and change the level for a validator on the fly.

Valid arguments are 'Default', 'Warnings', or 'Errors'.
Other values are ignored.